+++ src/App.css
@@ -50,3 +50,6 @@
   font-style: normal;
   font-display: block;
 }
+.scroll::-webkit-scrollbar {
+  display: none;
+}
