+++ src/components/TodoDone.jsx
@@ -1,10 +1,9 @@
 import React, { useState, useEffect } from "react";
 import styled from "styled-components";
-import TodoListItem from './todoListItem';
+import TodoListItem from './TodoListItem';
 
-
-const TodoListSytle = styled.div`
-    .todoListContainer{
+const TodoDoneSytle=styled.div`
+    .todoDoneContainer{
         position:relative;
         background-color: #262626;
         border-radius: 8px;
@@ -12,41 +11,56 @@ const TodoListSytle = styled.div`
         display:flex;
         flex-direction:column;
         flex:1;
-        padding:20px;
+        padding:0px 20px 20px 20px;
         align-items:center;
         margin:20px;
-      
+        height:800px;
+        overflow-y:scroll;
     }
-
-    .todoListTitle{
+    .todoDoneContainer::-webkit-scrollbar {
+        display: none;
+      }
+    .todoDoneTitle{
+        position:sticky;
+        top:0px;
+        background-color: #262626;
+        z-index:1;
         color:white;
         font-size:37px;
         font-weight:700;
-
         text-align: left;
-        width:90%;
+        width:100%;
+        padding:20px;
+        margin-bottom:5px;
+
     }
     .todoListPlus{
         position:absolute;
         width:26px;
         height:26px;
-        background-color:white;
+        backgroud-color:white;
         left:50%;
         bottom:-13px;
         border-radius:50%;
-     
 
     }
     .todoListItem{
         position:relative;
         display:flex;
         flex-direction:column;
-        margin:10px;
+        margin:15px;
         padding:10px;
         border-radius:8px;
         background-color:#D93535;
         width:90%;
+        color:white;
         
+    }
+    .todoListItemTitle{
+        margin:10px;
+    }
+    .todoListItemContent{
+        margin:10px;
     }
     .todoListItemXBtn{
         position:absolute;
@@ -72,42 +86,65 @@ const TodoListSytle = styled.div`
     }
     .todoListEditTitle{
         border-radius:8px;
-        background-color: rgb(200, 200, 200);
+        border:1px white solid;
+        background-color: rgb(200, 200, 200,0.5);
         margin:2px;
         font-size:20px;
         width:50%;
+        color:white;
+        font-style:bold;
+    }
+    .todoListEditTitle::placeholder{
+        color:white;
+        font-style:bold;
     }
     .todoListEditContent{
         border-radius:8px;
-        background-color: rgb(200, 200, 200);
+        border:1px white solid;
+        background-color: rgb(200, 200, 200,0.5);        
         margin:2px;
+        color:white;
+    }
+    .todoListEditContent::placeholder{
+        color:white;
+
     }
     .todoListEditBtn{
         border-radius:8px;
-        background-color: rgb(200, 200, 200);
+        border:1px white solid;
+        background-color: rgb(200, 200, 200,0.5);
         margin:2px;
+        width:50%;
+        color:white;
 
     }
     .hoverMoveTodoBtn{
-        display:none;
+        top:-20px;
+        position:absolute;
         border-radius:8px;
         background-color:white;
     }
     .hoverMoveProgressBtn{
         border-radius:8px;
+        position:absolute;
+        top:-20px;
+        left:50%;
         backgroudn-color:white;
     }  
     .hoverMoveDoneBtn{
+        display:none;
         border-radius:8px;
+        position:absolute;
         background-color:white;
     }
+    
     .todoListRedBtn{
         position:absolute;
         border-radius:50%;
         background-color:#D93535;
         width:30px;
         height:30px;
-        top:0px;
+        bottom:-5px;
         left:55%;
         border:1px solid white;
     }
@@ -117,7 +154,7 @@ const TodoListSytle = styled.div`
         background-color:#307FE2;
         width:30px;
         height:30px;
-        top:0px;
+        bottom:-5px;
         left:67%;
         border:1px solid white;
     } .todoListGreenBtn{
@@ -126,7 +163,7 @@ const TodoListSytle = styled.div`
         background-color:#00B294;
         width:30px;
         height:30px;
-        top:0px;
+        bottom:-5px;
         left:79%;
         border:1px solid white;
     } .todoListPurpleBtn{
@@ -135,17 +172,17 @@ const TodoListSytle = styled.div`
         background-color:#7678D1;
         width:30px;
         height:30px;
-        top:0px;
+        bottom:-5px;
         left:91%;
         border:1px solid white;
     }
+
 `
 
-function TodoList() {
+function TodoDone(){
     const [todos, setTodos] = useState([]);
     const [todoDone, setDone] = useState([]);
-    const [todoProgress, setTodoProgress] = useState([]);
-
+    const [todoProgress, setProgress] = useState([]);
     useEffect(() => {
         const savedTodos = JSON.parse(localStorage.getItem('todos')) || [];
         const savedDone = JSON.parse(localStorage.getItem('todoDone')) || [];
@@ -154,75 +191,74 @@ function TodoList() {
             setDone(savedDone)
         }
         if(savedProgress.length>0){
-            setTodoProgress(savedProgress)
+            setProgress(savedProgress)
         }
         if(savedTodos.length>0){
             setTodos(savedTodos);
         }
-    }, [todos]);
+    }, [todoDone]);
 
     const removeTodo = (index) => {
-        const updatedTodos = [...todos];
-        updatedTodos.splice(index, 1);
-        localStorage.setItem('todos', JSON.stringify(updatedTodos));
-        setTodos(updatedTodos);
+        const updatedDones = [...todoDone];
+        updatedDones.splice(index, 1);
+        localStorage.setItem('todoDone', JSON.stringify(updatedDones));
+        setDone(updatedDones);
     };
     const updateTodo = (index, updatedTitle, updatedContent) => {
-        const updatedTodos = [...todos];
-        updatedTodos[index] = { title: updatedTitle, content: updatedContent ,isEdit:false};
+        const updatedDone = [...todoDone];
+        updatedDone[index] = { title: updatedTitle, content: updatedContent ,isEdit:false};
 
-        localStorage.setItem('todos', JSON.stringify(updatedTodos));
-        setTodos(updatedTodos);
+        localStorage.setItem('todoDone', JSON.stringify(updatedDone));
+        setDone(updatedDone);
     };
     const addTodo = () => {
-        const newTodo = { title: '새로운 할일', content: '해야 할 일을 적어주세요' , isEdit:true};
-        const updatedTodos = [...todos, newTodo];
+        const newDone = { title: '새로운 할일', content: '해야 할 일을 적어주세요' , isEdit:true};
+        const updatedDones = [...todoDone, newDone];
 
-        localStorage.setItem('todos', JSON.stringify(updatedTodos));
-        setTodos(updatedTodos);
+        localStorage.setItem('todoDone', JSON.stringify(updatedDones));
+        setDone(updatedDones);
         
     };
     const moveToProgress=(index)=>{
-        const newProgress = { title: todos[index].title, content:todos[index].content , isEdit:false};
+        const newProgress = { title: todoDone[index].title, content:todoDone[index].content , isEdit:false};
         const updatedProgress = [...todoProgress, newProgress];
         localStorage.setItem('todoProgress', JSON.stringify(updatedProgress));
-        setTodoProgress(updatedProgress);
-        const updatedTodos = [...todos];
-        updatedTodos.splice(index, 1);
-        localStorage.setItem('todos', JSON.stringify(updatedTodos));
-        setTodos(updatedTodos);
-        
+        setProgress(updatedProgress);
+        const updatedDones = [...todoDone];
+        updatedDones.splice(index, 1);
+        localStorage.setItem('todoDone', JSON.stringify(updatedDones));
+        setDone(updatedDones);
+  
     };
-    const moveToDone=(index)=>{
-        const newDone = { title: todos[index].title, content:todos[index].content , isEdit:false};
-        const updatedDone = [...todoDone, newDone];
-        localStorage.setItem('todoDone', JSON.stringify(updatedDone));
-        setDone(updatedDone);
-        const updatedTodos = [...todos];
-        updatedTodos.splice(index, 1);
-        localStorage.setItem('todos', JSON.stringify(updatedTodos));
-        setTodos(updatedTodos);
+    const moveToTodo=(index)=>{
+        const newTodo = { title: todoDone[index].title, content:todoDone[index].content , isEdit:false};
+        const updatedTodo = [...todos, newTodo];
+        localStorage.setItem('todos', JSON.stringify(updatedTodo));
+        setDone(updatedTodo);
+        const updatedDones = [...todoDone];
+        updatedDones.splice(index, 1);
+        localStorage.setItem('todoDone', JSON.stringify(updatedDones));
+        setDone(updatedDones);
+
     };
-    return (
-        <TodoListSytle>
-            <div className="todoListContainer">
-                <div className="todoListTitle">
-                    To Do 🐣
+    return(
+        <TodoDoneSytle>
+            <div className="todoDoneContainer">
+                <div className="todoDoneTitle"> 
+                    Done 🕊
                 </div>
-                {todos.map((todo, index) => (
+                {todoDone.map((todo, index) => (
                     <TodoListItem key={index} title={todo.title} content={todo.content}
                         onRemove={() => removeTodo(index)}
                         onUpdate={(updatedTitle, updatedContent) => updateTodo(index, updatedTitle, updatedContent)
                         }isEdit={todo.isEdit}
-                        onMoveProgress={()=>moveToProgress(index)}
-                        onMoveDone={()=>moveToDone(index)} />
+                        onMoveTodo={()=>moveToTodo(index)}
+                        onMoveProgress={()=>moveToProgress(index)} />
                 ))}
                 <button className="todoListPlus" onClick={addTodo}> +</button>
+
             </div>
-                    
-        </TodoListSytle>
+        </TodoDoneSytle>
     )
 }
-
-
-export default TodoList;
+export default TodoDone;
