+++ src/components/TodoList.jsx
@@ -1,9 +1,10 @@
 import React, { useState, useEffect } from "react";
 import styled from "styled-components";
-import TodoListItem from './todoListItem';
+import TodoListItem from './TodoListItem';
 
-const TodoProgressSytle=styled.div`
-    .todoProgressContainer{
+
+const TodoListSytle = styled.div`
+    .todoListContainer{
         position:relative;
         background-color: #262626;
         border-radius: 8px;
@@ -11,39 +12,57 @@ const TodoProgressSytle=styled.div`
         display:flex;
         flex-direction:column;
         flex:1;
-        padding:20px;
+        padding:0px 20px 20px 20px;
         align-items:center;
         margin:20px;
+        height:800px;
+        overflow-y:auto;
+      
     }
-    .todoProgressTitle{
+    .todoListContainer::-webkit-scrollbar {
+        display: none;
+      }
+    .todoListTitle{
+        position:sticky;
+        top:0px;
+        background-color: #262626;
+        z-index:1;
         color:white;
-        font-weight:700;
-
         font-size:37px;
+        font-weight:700;
         text-align: left;
-        width:90%;
+        width:100%;
+        padding:20px;
+        margin-bottom:5px;
 
     }
     .todoListPlus{
         position:absolute;
         width:26px;
         height:26px;
-        backgroud-color:white;
+        background-color:white;
         left:50%;
         bottom:-13px;
         border-radius:50%;
-
+    
     }
     .todoListItem{
         position:relative;
         display:flex;
         flex-direction:column;
-        margin:10px;
+        margin:15px;
         padding:10px;
         border-radius:8px;
         background-color:#D93535;
         width:90%;
-        
+        color:white;
+
+    }
+    .todoListItemTitle{
+        margin:10px;
+    }
+    .todoListItemContent{
+        margin:10px;
     }
     .todoListItemXBtn{
         position:absolute;
@@ -67,44 +86,69 @@ const TodoProgressSytle=styled.div`
         flex-direction:column;
         position:relative;
     }
+    
+
     .todoListEditTitle{
         border-radius:8px;
-        background-color: rgb(200, 200, 200);
+        border:1px white solid;
+        background-color: rgb(200, 200, 200,0.5);
         margin:2px;
         font-size:20px;
         width:50%;
+        color:white;
+        font-style:bold;
+    }
+    .todoListEditTitle::placeholder{
+        color:white;
+        font-style:bold;
     }
     .todoListEditContent{
         border-radius:8px;
-        background-color: rgb(200, 200, 200);
+        border:1px white solid;
+        background-color: rgb(200, 200, 200,0.5);        
         margin:2px;
+        color:white;
+    }
+    .todoListEditContent::placeholder{
+        color:white;
+
     }
     .todoListEditBtn{
         border-radius:8px;
-        background-color: rgb(200, 200, 200);
+        border:1px white solid;
+        background-color: rgb(200, 200, 200,0.5);
         margin:2px;
+        width:50%;
+        color:white;
 
     }
     .hoverMoveTodoBtn{
+        display:none;
+        position:absolute;
         border-radius:8px;
         background-color:white;
     }
     .hoverMoveProgressBtn{
-        display:none;
         border-radius:8px;
+        position:absolute;
+        top:-20px;
         backgroudn-color:white;
     }  
     .hoverMoveDoneBtn{
         border-radius:8px;
+        position:absolute;
+        top:-20px;
+        left:50%;
         background-color:white;
     }
+    
     .todoListRedBtn{
         position:absolute;
         border-radius:50%;
         background-color:#D93535;
         width:30px;
         height:30px;
-        top:0px;
+        bottom:-5px;
         left:55%;
         border:1px solid white;
     }
@@ -114,7 +158,7 @@ const TodoProgressSytle=styled.div`
         background-color:#307FE2;
         width:30px;
         height:30px;
-        top:0px;
+        bottom:-5px;
         left:67%;
         border:1px solid white;
     } .todoListGreenBtn{
@@ -123,7 +167,7 @@ const TodoProgressSytle=styled.div`
         background-color:#00B294;
         width:30px;
         height:30px;
-        top:0px;
+        bottom:-5px;
         left:79%;
         border:1px solid white;
     } .todoListPurpleBtn{
@@ -132,16 +176,17 @@ const TodoProgressSytle=styled.div`
         background-color:#7678D1;
         width:30px;
         height:30px;
-        top:0px;
+        bottom:-5px;
         left:91%;
         border:1px solid white;
     }
 `
 
-function TodoProgress(){
+function TodoList() {
     const [todos, setTodos] = useState([]);
     const [todoDone, setDone] = useState([]);
-    const [todoProgress, setProgress] = useState([]);
+    const [todoProgress, setTodoProgress] = useState([]);
+
     useEffect(() => {
         const savedTodos = JSON.parse(localStorage.getItem('todos')) || [];
         const savedDone = JSON.parse(localStorage.getItem('todoDone')) || [];
@@ -150,76 +195,73 @@ function TodoProgress(){
             setDone(savedDone)
         }
         if(savedProgress.length>0){
-            setProgress(savedProgress)
+            setTodoProgress(savedProgress)
         }
         if(savedTodos.length>0){
             setTodos(savedTodos);
         }
-    }, [todoProgress]);
+    }, [todos]);
 
     const removeTodo = (index) => {
-        const updatedProgress = [...todoProgress];
-        updatedProgress.splice(index, 1);
-        localStorage.setItem('todoProgress', JSON.stringify(updatedProgress));
-        setProgress(updatedProgress);
+        const updatedTodos = [...todos];
+        updatedTodos.splice(index, 1);
+        localStorage.setItem('todos', JSON.stringify(updatedTodos));
+        setTodos(updatedTodos);
     };
     const updateTodo = (index, updatedTitle, updatedContent) => {
-        const updatedProgress = [...todoProgress];
-        updatedProgress[index] = { title: updatedTitle, content: updatedContent ,isEdit:false};
+        const updatedTodos = [...todos];
+        updatedTodos[index] = { title: updatedTitle, content: updatedContent ,isEdit:false};
 
-        localStorage.setItem('todoProgress', JSON.stringify(updatedProgress));
-        setProgress(updatedProgress);
-
+        localStorage.setItem('todos', JSON.stringify(updatedTodos));
+        setTodos(updatedTodos);
     };
     const addTodo = () => {
-        const newProgress = { title: '새로운 할일', content: '해야 할 일을 적어주세요' , isEdit:true};
+        const newTodo = { title: '', content: '' , isEdit:true};
+        const updatedTodos = [...todos, newTodo];
+        localStorage.setItem('todos', JSON.stringify(updatedTodos));
+        setTodos(updatedTodos);
+    };
+    const moveToProgress=(index)=>{
+        const newProgress = { title: todos[index].title, content:todos[index].content , isEdit:false};
         const updatedProgress = [...todoProgress, newProgress];
-
         localStorage.setItem('todoProgress', JSON.stringify(updatedProgress));
-        setProgress(updatedProgress);
-
+        setTodoProgress(updatedProgress);
+        const updatedTodos = [...todos];
+        updatedTodos.splice(index, 1);
+        localStorage.setItem('todos', JSON.stringify(updatedTodos));
+        setTodos(updatedTodos);
         
     };
     const moveToDone=(index)=>{
-        const newDone = { title: todoProgress[index].title, content:todoProgress[index].content , isEdit:false};
+        const newDone = { title: todos[index].title, content:todos[index].content , isEdit:false};
         const updatedDone = [...todoDone, newDone];
         localStorage.setItem('todoDone', JSON.stringify(updatedDone));
         setDone(updatedDone);
-        const updatedProgress = [...todoProgress];
-        updatedProgress.splice(index, 1);
-        localStorage.setItem('todoProgress', JSON.stringify(updatedProgress));
-        setProgress(updatedProgress);
+        const updatedTodos = [...todos];
+        updatedTodos.splice(index, 1);
+        localStorage.setItem('todos', JSON.stringify(updatedTodos));
+        setTodos(updatedTodos);
     };
-    const moveToTodo=(index)=>{
-        const newTodo = { title: todoProgress[index].title, content:todoProgress[index].content , isEdit:false};
-        const updatedTodo = [...todos, newTodo];
-        localStorage.setItem('todos', JSON.stringify(updatedTodo));
-        setDone(updatedTodo);
-        const updatedProgress = [...todoProgress];
-        updatedProgress.splice(index, 1);
-        localStorage.setItem('todoProgress', JSON.stringify(updatedProgress));
-        setProgress(updatedProgress);
-
-    };
-    return(
-        <TodoProgressSytle>
-            <div className="todoProgressContainer">
-                <div className="todoProgressTitle"> 
-                    In Progress 🐥
+    return (
+        <TodoListSytle>
+            <div className="todoListContainer">
+                <div className="todoListTitle">
+                    To Do 🐣
                 </div>
-                {todoProgress.map((todo, index) => (
+                {todos.map((todo, index) => (
                     <TodoListItem key={index} title={todo.title} content={todo.content}
                         onRemove={() => removeTodo(index)}
                         onUpdate={(updatedTitle, updatedContent) => updateTodo(index, updatedTitle, updatedContent)
                         }isEdit={todo.isEdit}
-                        onMoveTodo={()=>moveToTodo(index)}
+                        onMoveProgress={()=>moveToProgress(index)}
                         onMoveDone={()=>moveToDone(index)} />
                 ))}
                 <button className="todoListPlus" onClick={addTodo}> +</button>
-
             </div>
-        </TodoProgressSytle>
+                    
+        </TodoListSytle>
     )
 }
 
-export default TodoProgress;
+
+export default TodoList;
