+++ src/components/TodoProgress.jsx
@@ -1,9 +1,9 @@
 import React, { useState, useEffect } from "react";
 import styled from "styled-components";
-import TodoListItem from './todoListItem';
+import TodoListItem from './TodoListItem';
 
-const TodoDoneSytle=styled.div`
-    .todoDoneContainer{
+const TodoProgressSytle=styled.div`
+    .todoProgressContainer{
         position:relative;
         background-color: #262626;
         border-radius: 8px;
@@ -11,17 +11,27 @@ const TodoDoneSytle=styled.div`
         display:flex;
         flex-direction:column;
         flex:1;
-        padding:20px;
+        padding:0px 20px 20px 20px;
         align-items:center;
         margin:20px;
+        height:800px;
+        overflow-y:auto;
     }
-    .todoDoneTitle{
+    .todoProgressContainer::-webkit-scrollbar {
+        display: none;
+      }
+    .todoProgressTitle{
+        position:sticky;
+        top:0px;
+        background-color: #262626;
+        z-index:1;
         color:white;
         font-size:37px;
         font-weight:700;
-
         text-align: left;
-        width:90%;
+        width:100%;
+        padding:20px;
+        margin-bottom:5px;
 
     }
     .todoListPlus{
@@ -38,12 +48,18 @@ const TodoDoneSytle=styled.div`
         position:relative;
         display:flex;
         flex-direction:column;
-        margin:10px;
+        margin:15px;
         padding:10px;
         border-radius:8px;
         background-color:#D93535;
         width:90%;
-        
+        color:white;
+    }
+    .todoListItemTitle{
+        margin:10px;
+    }
+    .todoListItemContent{
+        margin:10px;
     }
     .todoListItemXBtn{
         position:absolute;
@@ -69,42 +85,66 @@ const TodoDoneSytle=styled.div`
     }
     .todoListEditTitle{
         border-radius:8px;
-        background-color: rgb(200, 200, 200);
+        border:1px white solid;
+        background-color: rgb(200, 200, 200,0.5);
         margin:2px;
         font-size:20px;
         width:50%;
+        color:white;
+        font-style:bold;
+    }
+    .todoListEditTitle::placeholder{
+        color:white;
+        font-style:bold;
     }
     .todoListEditContent{
         border-radius:8px;
-        background-color: rgb(200, 200, 200);
+        border:1px white solid;
+        background-color: rgb(200, 200, 200,0.5);        
         margin:2px;
+        color:white;
+    }
+    .todoListEditContent::placeholder{
+        color:white;
+
     }
     .todoListEditBtn{
         border-radius:8px;
-        background-color: rgb(200, 200, 200);
+        border:1px white solid;
+        background-color: rgb(200, 200, 200,0.5);
         margin:2px;
+        width:50%;
+        color:white;
 
     }
     .hoverMoveTodoBtn{
+        top:-20px;
+        position:absolute;
         border-radius:8px;
         background-color:white;
     }
     .hoverMoveProgressBtn{
+        display:none;
         border-radius:8px;
+        position:absolute;
+        top:-20px;
         backgroudn-color:white;
     }  
     .hoverMoveDoneBtn{
-        display:none;
         border-radius:8px;
+        position:absolute;
+        top:-20px;
+        left:50%;
         background-color:white;
     }
+    
     .todoListRedBtn{
         position:absolute;
         border-radius:50%;
         background-color:#D93535;
         width:30px;
         height:30px;
-        top:0px;
+        bottom:-5px;
         left:55%;
         border:1px solid white;
     }
@@ -114,7 +154,7 @@ const TodoDoneSytle=styled.div`
         background-color:#307FE2;
         width:30px;
         height:30px;
-        top:0px;
+        bottom:-5px;
         left:67%;
         border:1px solid white;
     } .todoListGreenBtn{
@@ -123,7 +163,7 @@ const TodoDoneSytle=styled.div`
         background-color:#00B294;
         width:30px;
         height:30px;
-        top:0px;
+        bottom:-5px;
         left:79%;
         border:1px solid white;
     } .todoListPurpleBtn{
@@ -132,14 +172,13 @@ const TodoDoneSytle=styled.div`
         background-color:#7678D1;
         width:30px;
         height:30px;
-        top:0px;
+        bottom:-5px;
         left:91%;
         border:1px solid white;
     }
-
 `
 
-function TodoDone(){
+function TodoProgress(){
     const [todos, setTodos] = useState([]);
     const [todoDone, setDone] = useState([]);
     const [todoProgress, setProgress] = useState([]);
@@ -156,69 +195,71 @@ function TodoDone(){
         if(savedTodos.length>0){
             setTodos(savedTodos);
         }
-    }, [todoDone]);
+    }, [todoProgress]);
 
     const removeTodo = (index) => {
-        const updatedDones = [...todoDone];
-        updatedDones.splice(index, 1);
-        localStorage.setItem('todoDone', JSON.stringify(updatedDones));
-        setDone(updatedDones);
-    };
-    const updateTodo = (index, updatedTitle, updatedContent) => {
-        const updatedDone = [...todoDone];
-        updatedDone[index] = { title: updatedTitle, content: updatedContent ,isEdit:false};
-
-        localStorage.setItem('todoDone', JSON.stringify(updatedDone));
-        setDone(updatedDone);
-    };
-    const addTodo = () => {
-        const newDone = { title: '새로운 할일', content: '해야 할 일을 적어주세요' , isEdit:true};
-        const updatedDones = [...todoDone, newDone];
-
-        localStorage.setItem('todoDone', JSON.stringify(updatedDones));
-        setDone(updatedDones);
-        
-    };
-    const moveToProgress=(index)=>{
-        const newProgress = { title: todoDone[index].title, content:todoDone[index].content , isEdit:false};
-        const updatedProgress = [...todoProgress, newProgress];
+        const updatedProgress = [...todoProgress];
+        updatedProgress.splice(index, 1);
         localStorage.setItem('todoProgress', JSON.stringify(updatedProgress));
         setProgress(updatedProgress);
-        const updatedDones = [...todoDone];
-        updatedDones.splice(index, 1);
-        localStorage.setItem('todoDone', JSON.stringify(updatedDones));
-        setDone(updatedDones);
-  
+    };
+    const updateTodo = (index, updatedTitle, updatedContent) => {
+        const updatedProgress = [...todoProgress];
+        updatedProgress[index] = { title: updatedTitle, content: updatedContent ,isEdit:false};
+
+        localStorage.setItem('todoProgress', JSON.stringify(updatedProgress));
+        setProgress(updatedProgress);
+
+    };
+    const addTodo = () => {
+        const newProgress = { title: '새로운 할일', content: '해야 할 일을 적어주세요' , isEdit:true};
+        const updatedProgress = [...todoProgress, newProgress];
+
+        localStorage.setItem('todoProgress', JSON.stringify(updatedProgress));
+        setProgress(updatedProgress);
+
+        
+    };
+    const moveToDone=(index)=>{
+        const newDone = { title: todoProgress[index].title, content:todoProgress[index].content , isEdit:false};
+        const updatedDone = [...todoDone, newDone];
+        localStorage.setItem('todoDone', JSON.stringify(updatedDone));
+        setDone(updatedDone);
+        const updatedProgress = [...todoProgress];
+        updatedProgress.splice(index, 1);
+        localStorage.setItem('todoProgress', JSON.stringify(updatedProgress));
+        setProgress(updatedProgress);
     };
     const moveToTodo=(index)=>{
-        const newTodo = { title: todoDone[index].title, content:todoDone[index].content , isEdit:false};
+        const newTodo = { title: todoProgress[index].title, content:todoProgress[index].content , isEdit:false};
         const updatedTodo = [...todos, newTodo];
         localStorage.setItem('todos', JSON.stringify(updatedTodo));
         setDone(updatedTodo);
-        const updatedDones = [...todoDone];
-        updatedDones.splice(index, 1);
-        localStorage.setItem('todoDone', JSON.stringify(updatedDones));
-        setDone(updatedDones);
+        const updatedProgress = [...todoProgress];
+        updatedProgress.splice(index, 1);
+        localStorage.setItem('todoProgress', JSON.stringify(updatedProgress));
+        setProgress(updatedProgress);
 
     };
     return(
-        <TodoDoneSytle>
-            <div className="todoDoneContainer">
-                <div className="todoDoneTitle"> 
-                    Done 🕊
+        <TodoProgressSytle>
+            <div className="todoProgressContainer">
+                <div className="todoProgressTitle"> 
+                    In Progress 🐥
                 </div>
-                {todoDone.map((todo, index) => (
+                {todoProgress.map((todo, index) => (
                     <TodoListItem key={index} title={todo.title} content={todo.content}
                         onRemove={() => removeTodo(index)}
                         onUpdate={(updatedTitle, updatedContent) => updateTodo(index, updatedTitle, updatedContent)
                         }isEdit={todo.isEdit}
                         onMoveTodo={()=>moveToTodo(index)}
-                        onMoveProgress={()=>moveToProgress(index)} />
+                        onMoveDone={()=>moveToDone(index)} />
                 ))}
                 <button className="todoListPlus" onClick={addTodo}> +</button>
 
             </div>
-        </TodoDoneSytle>
+        </TodoProgressSytle>
     )
 }
-export default TodoDone;
+
+export default TodoProgress;
